+++ crowdsec.php
@@ -4,7 +4,7 @@
  * Plugin URI: https://github.com/crowdsecurity/cs-wordpress-bouncer
  * Description: Safer Together. Protect your WordPress application with CrowdSec.
  * Tags: security, captcha, ip-blocker, crowdsec, hacker-protection, appsec
- * Version: 2.9.0
+ * Version: 2.10.0
  * Author: CrowdSec
  * Author URI: https://www.crowdsec.net/
  * Github: https://github.com/crowdsecurity/cs-wordpress-bouncer
@@ -12,8 +12,8 @@
  * License URI: https://opensource.org/licenses/MIT
  * Requires PHP: 7.2
  * Requires at least: 4.9
- * Tested up to: 6.7
- * Stable tag: 2.9.0
+ * Tested up to: 6.8
+ * Stable tag: 2.10.0
  * Text Domain: crowdsec-wp
  * First release: 2021.
  */